refactor(task): move Task entity and repository into task/ package
Start of the DDD migration for Task (see backend/TASK_MIGRATION_FINDINGS.md).
Moves Task and TaskRepository from the layer-based domain package into the
new package-by-feature task/ package, mirroring category/. Notification
(kept in domain/ for now) is updated to import Task from its new location.
