+++ backend/src/main/java/br/com/calendar/domain/Notification.java
@@ -1,6 +1,7 @@
 package br.com.calendar.domain;
 
 import br.com.calendar.common.BaseEntity;
+import br.com.calendar.task.Task;
 import br.com.calendar.user.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
