+++ backend/src/main/java/br/com/calendar/task/Task.java
@@ -1,4 +1,4 @@
-package br.com.calendar.domain;
+package br.com.calendar.task;
 
 import br.com.calendar.common.BaseEntity;
 import br.com.calendar.category.Category;
